# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GDS Preview API",
    "description": "Creates live preview deployments of React apps built with GDS design systems (Gofore GDS or GDS-VERO). Send React component code, a designSystem id, and receive a unique public preview URL.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "https://gds-preview-app-muxj.vercel.app",
      "description": "Production"
    }
  ],
  "paths": {
    "/api/create-preview": {
      "post": {
        "operationId": "createPreview",
        "summary": "Create a GDS preview deployment",
        "description": "Receives React/TSX code for App.tsx (and optional extra component files), wraps it in a Vite + GDS template for the selected design system, and deploys it to Vercel. Returns a unique preview URL. Each deployment is isolated — previews do not overwrite each other.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["appCode", "designSystem"],
                "properties": {
                  "designSystem": {
                    "type": "string",
                    "enum": ["gds", "gds-vero"],
                    "description": "Which design system stack to use. Use \"gds\" for Gofore GDS (@gdesignsystem/*). Use \"gds-vero\" for Verohallinto / vero.fi GDS-VERO (@gds-vero/*). Required so multiple Custom GPTs can share this API safely."
                  },
                  "appCode": {
                    "type": "string",
                    "description": "The full contents of src/App.tsx. Must export a default React component. GDSProvider is already wrapped in main.tsx — do not add it in App.tsx."
                  },
                  "extraFiles": {
                    "type": "object",
                    "description": "Optional additional source files. Keys must start with 'src/' (e.g. 'src/components/Header.tsx'). Values are the full file contents as strings.",
                    "additionalProperties": {
                      "type": "string"
                    }
                  },
                  "title": {
                    "type": "string",
                    "description": "Optional page title for the HTML document. Defaults to 'GDS Preview'."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deployment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "previewUrl": {
                      "type": "string",
                      "description": "URL to share with the user. Opens a building page and redirects to the live preview when ready."
                    },
                    "finalPreviewUrl": {
                      "type": "string",
                      "description": "The final Vercel preview domain (redirect target after build)"
                    },
                    "deploymentId": {
                      "type": "string",
                      "description": "Vercel deployment ID"
                    },
                    "previewId": {
                      "type": "string",
                      "description": "Short unique ID for this preview"
                    },
                    "designSystem": {
                      "type": "string",
                      "description": "The design system used for this deployment"
                    },
                    "status": {
                      "type": "string",
                      "description": "Deployment status (initially 'building')"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (missing appCode, wrong designSystem, cross-stack imports, blocked code patterns)"
          },
          "401": {
            "description": "Unauthorized (missing or invalid API secret)"
          },
          "502": {
            "description": "Vercel deployment failed"
          }
        }
      }
    }
  }
}